extract metablock into own file
